northd: Fix virtual port ip advertisment via BGP.
When dynamic-routing-redistribute option is set to connected-as-host for
logical_router_port where dynamic-routing is enabled, virtual port ips
are advertise without taking into account if the port is bound locally.
Fix the issue setting the advertised route tracked_port to the virtual
parent port in order to allow the CMS to mange the host route priority
or to not advertised the host route if the port is not local.

Reported-at: https://issues.redhat.com/browse/FDP-1940
Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit 0bf8aec8c178742fcbe18d78e228ae1aef475521)