controller: Prevent other chassis with same encap IP.
If you set a chassis' encap ip to be the same as that of another
chassis, OVN should detect it, log an error, and delete the chassis
record.

But this was not the case; OVN would allow the transaction. However, the
commit to the southbound database would fail, because there was already
an encap record with that same ip and type (from the other chassis).
The OVS db on the problematic chassis would then rapidly increase in
size. Essentially, the controller would get stuck in a loop of deleting
and recreating the tunnels.

This change makes it so when encaps are reconfigured, before being
built, in chassis_build_encaps(), the controller detects if there is
already an existing encap tunnel of the same ip and type. If so, and
there are no good (unique) encaps configured to use, only bad ones (as
in already used on another chassis), the controller logs an error,
refuses to build the encaps, and deletes the associated chassis and
chassis private record. If there are both good encaps IPs and bad ones
configured, the good ones will be used to build encaps and the bad ones
will be ignored.

In chassis_build_encaps(), only if an Encap record does not already
exist will a new record be built. And options will only be updated if
they've changed.

Another change this patch introduces is if there's currently a
southbound transaction in flight, encaps_run() will not run. This is to
maintain symmetry with chassis_run()'s behavior. This change prevents an
inconsistency where the controller saw ports in the local OVS db but not
in sb-db, and then deleted and recreated them.

Reported-at: https://issues.redhat.com/browse/FDP-2013
Co-authored-by: Nicholas Hubbard <nicholashubbard@posteo.net>
Signed-off-by: Nicholas Hubbard <nicholashubbard@posteo.net>
Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit c2dd8593c8e81ed1378c086e5f502b4f35472e79)