system-test: Make sure we can bind the virtual port right away.
The dynamic-routing test was rebinding virtual port almost back to
back. That could lead to a situation when the virtual port wouldn't
be bound because of the backoff. Make sure the backoff is removed
before we attempt to bind it again.

Fixes: 0bf8aec8c178 ("northd: Fix virtual port ip advertisment via BGP.")
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit b362efc5eb70196b7dd84cbadf3698acb6534a92)