tests: Collect the relevant artifacts after each multinode test.
We would lose all logs and databases once the boundary between
tests is crossed, that's by design so other tests don't start with
unwanted data. However, in case of test failure we want to preserve
the data to help with the analysis of the failure. Make sure we
collect databases and logs after each test.

Reported-at: https://issues.redhat.com/browse/FDP-2133
Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>