tests: Make the ACL Sampling test more reliable.
The test was using client and server to generate traffic that would
be sampled. The sampling check relied on the fact that the last
echo from server would arrive when the client is already closed.
That wasn't always the case which led to missing samples for ICMP
error. Make sure the server waits a little before sending the last
packet, giving the client a chance to properly exit first. Also make
the delay between the two packets sent by client smaller to slightly
shorten the time required to execute the test.

Reported-at: https://issues.redhat.com/browse/FDP-1941
Acked-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>