ofctrl: Prevent duplicate desired flow references.
Address sets may contain duplicate values such as "::1" and "::01" or
"1.1.1.1" and "1.1.1.1/32", which produce the same desired flow.
When adding such a duplicate to an existing address set, an lflow using
the address set that is updated incrementally will produce the same
desired flow that it produced for the original value, find that this
desired flow already exist and link the lflow to the desired flow again.
This currently assumes that the existing desired flow was produced by
another lflow, which is not the case here and results in a second
reference from the lflow to the desired flow.

This then leads to a crash later on when removing or recomputing the
lflow and going through remove_flows_from_sb_to_flow by tripping over
the ovs_assert(ovs_list_is_empty(&f->list_node)) as the code there does
not anticipate seeing the same desired flow twice for one lflow.

Fix this by detecting and preventing the duplicate reference. Also log
a warning as this likely indicates a problem somewhere else.

Fixes: 7c0b538762f6 ("ovn-controller: Handle addresses addition in address set incrementally.")
Signed-off-by: Felix Moebius <felix.moebius@mail.schwarz>
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit 1f6866060cbfd5a2bf2e8523f6e5e50b2ea2e6c7)