en-datapath-sync: Avoid unnecessary datapath recomputes.
In OVN 25.09, we updated the southbound database's Datapath_Binding
table to contain a new nb_uuid column that contains the corresponding
Logical_Switch or Logical_Router UUID. For safety's sake, we created an
accessor function to get the NB UUID from a SB Datapath_Binding. This
made it so that on upgrade to OVN 25.09, we could find the NB UUID the
new way (through the nb_uuid column) or the old way (through
external-ids fields).

In the datapath sync code, we call this accessor function in
find_synced_datapath_from_sb() in order to get the NB UUID and hash that
UUID. However, once we started iterating through the hashmap of synced
datapaths, we reverted to referencing the nb_uuid field directly. This
means that on upgrades, this nb_uuid field will not have a UUID. This
means we won't find a corresponding synced datapath and will have to
recompute.

This commit fixes the problem by using the nb_uuid we looked up in the
accessor function. This will reduce the risk of unnecessary recomputes.

Fixes: 5d289d9acad8 ("ovn-sb: Adjust Datapath_Binding schema to avoid problems.")
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>