pinctrl: Make sure we can learn IGMP groups only on switches.
The IGMP group could be accidentally learned on LR datapath instead
of LS. While the IGMP Group would be removed by northd if it was the
case this still creates some churn in SB database. This could happen
under the following conditions:

1) pinctrl thread will receive packet and stored IGMP group candidate
   in hmap, with datapath tunnel_key as key.
2) The LS datapath tunnel_key is reused by LR.
3) ovn-controller will process those changes and will run
   ip_mcast_sync().
4) The ip_mcast_sync() will walk through IGMP group candidates
   retrieving datapath for given tunnel_key being LR at this point.
5) The IGMP group is stored in SB with a reference to LR.

There is one more condition, the pinctrl thread can't run after
2) otherwise it would update the in memory ip_mcast_snoop map.

To prevent that make sure we check if the retrieved datapath
is LS and skip the processing if that's not the case.

Fixes: 70ff8243040f ("OVN: Add IGMP SB definitions and ovn-controller support")
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>