lflows: Redefine ovn_lflow_add_with_dp_group().
This redefines the ovn_lflow_add_with_dp_group() macro to use the
struct-based ovn_lflow_add() function. This means the mandatory hint
argument is now replaced using the WITH_HINT macro.

This macro is not being defined in terms of ovn_lflow_add() since
ovn_lflow_add_requires an ovn_datapath struct, and in this case, the
argument is unnecessary.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>