ovn-ic: Introduce incremental processing.
Initial implementation adds a single node (ic). This single
node executes the ovn-ic processing pipeline but does not do so
incrementally.

In order to develop incremental processing for ovn-ic, the code
will be organised with a .c/.h file for each I-P node following
the naming convention en-<node name>.c/.h. These files will
contain definition of the node data, the main node processing
functions and change handlers (if any). The purpose of these nodes
will be coordination of the nodes work and implemention of the
relevant interfaces to plug into the I-P framework. The actual
work that will be executed by the node will be organised into
a companion file or files. Ideally this file will follow the
naming convention of the node: e.g. en-<node name>.c is
associated with <node name>.c.

Initial node topology sees the ic node dependent on all DB
nodes. This structure and approach was heavily influenced
by the way I+P was developed and evolved in the northd daemon.

Signed-off-by: Paulo Guilherme Silva <guilherme.paulo@luizalabs.com>
Signed-off-by: Numan Siddique <numans@ovn.org>