ovn-ic: Support dumping a visual representation of the I-P graph.
This commit updates the ability to dump a (partial) representation
of the incremental processing graph. Now It's currently supported for
ovn-ic too through a command-line argument, --dump-inc-proc-graph[=<i-p-node>].

If the command line argument is present the binary (ovn-ic)
just dumps a representation of the incremental
processing graph in DOT format to stdout.  The binary then exits.

If the '<i-p-node>' optional argument value is present the
representation only includes nodes up to (and including) '<i-p-node>'.

Signed-off-by: Paulo Guilherme Silva <guilherme.paulo@luizalabs.com>
Signed-off-by: Numan Siddique <numans@ovn.org>