lflows: Avoid unnecesary datapath validity checks.
When the search mode is LFLOW_TABLE_SEARCH_SBUUID, we start the syncing
process by iterating over all in-memory logical flows and syncing those
with the southbound database. We then iterate over the southbound
database entries. If the southbound entry does not match one of the
in-memory logical flows we just synced, then we delete the southbound
Logical flow. If the southbound entry matches on one of the
in-memory logical flows that we just synced, then we perform a check to
ensure that the southbound logical flow refers to valid datapaths. If
the datapath is invalid, then we delete the southbound logical flow.

This datapath check is not necessary. We started the process by syncing
using the in-memory logical flows. By definition, these logical flows can
only be created for synced datapaths, which are valid. Since the in-memory
logical flow has already been synced with the southbound logical flow,
it means that the southbound logical flow has valid datapaths.

This patch implements this change by only performing datapath validity
checks when the search mode is LFLOW_TABLE_SEARCH_FIELDS. The datapath
validity check is still necessary in this case, since we need the
datapath in order build an enum ovn_stage to help search for the
matching in-memory logical flow.

Fixes: 674738fc9806 ("northd: Refactor lflow table sync to sb.")
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>