controller: Add missing nw_tll field to match against legit NAs
When an ICMPv6 packet with icmp_type=128 and icmp_code=0 traverses
the OFTABLE_CHK_IN_PORT_SEC_ND table, the nd_target and nd_ttl fields
will be included in the megaflow. This causes the revalidator to
invalidate the flow, which in turn triggers an upcall. The root cause
is that the final-stage map is included in the wildcards during
classifier rule lookup.

By adding the missing nw_ttl field to the OpenFlow rule, the previous
three stage maps-metadata, L2, and L3-are only included in the
wildcards. As a result, the nd_target and nd_ttl fields are excluded
from the megaflow, preventing unnecessary revalidation and upcalls.

Fixes: 8cab00b ("ovn-controller: Add OF rules for port security.")
Reported-at: https://mail.openvswitch.org/pipermail/ovs-discuss/2025-December/053889.html
Signed-off-by: Jun Gu <jun.gu@easystack.cn>