br-controller: Fix probable NULL deferencing when encoding actions.
Since we use the common action parsing code between ovn-controller
and br-controller, ovnacts_encode() expects to pass a
group table and meter table for the actions - ct_lb, select, log and
set_meter.  Until we have a separate action parsing definitions
for br-controller, we just maintain these tables and pass it in
the 'struct ovnact_encode_params' in the ovnact_encode() to avoid
probable NULL dereferncing.

It is better to address this way, than checking for
group_table/meter_table NULL in the encode functions.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Ales Musil <amusil@redhat.com>